+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changes
 
 * 1.1.next in progress
+  * Fix #69 by excluding `:project` from basis if no `deps.edn` exists in the current directory.
   * Address #67 by annotating most namespaces with `^:no-doc`.
   * Update `polylith` template to follow current Polylith practices.